Add condition to check if user has enabled data removal when uninstalling plugin

+++ plugin/uninstall.php
@@ -3,6 +3,11 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
     exit;
 }
 
-global $wpdb;
+// Check if the user has selected removing all plugin data from the database after uninstalling the plugin - defaults to 0 (false)
+$remove_data = get_option( 'basic_auth_plugin_remove_data_after_uninstall', '0' );
 
-$wpdb->query("DELETE FROM {$wpdb->options} WHERE option_name LIKE 'basic_auth_plugin_%'");
+if ( $remove_data === '1' ) {
+  global $wpdb;
+
+  $wpdb->query("DELETE FROM {$wpdb->options} WHERE option_name LIKE 'basic_auth_plugin_%'");
+}
